youtube_summarizer/Backend/requirements.txt
# Web framework
fastapi>=0.111.0
uvicorn[standard]>=0.29.0
python-multipart>=0.0.9

# Data validation
pydantic>=2.7.0

# Environment variables
python-dotenv>=1.0.0

# YouTube transcript extraction
youtube-transcript-api>=0.6.2

# Google Gemini AI
google-generativeai>=0.7.0

fastapi>=0.110.0
uvicorn[standard]>=0.29.0
python-dotenv>=1.0.0
google-generativeai>=0.7.0
youtube-transcript-api>=0.6.2
pydantic>=2.0.0